Reorganise parser tests
