+++ community/cypher/src/test/scala/org/neo4j/cypher/internal/parser/legacy/ParserTest.scala
@@ -17,12 +17,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.parser
+package org.neo4j.cypher.internal.parser.legacy
 
 import org.scalatest.Assertions
 import util.parsing.input.CharSequenceReader
 import scala.util.parsing.combinator.JavaTokenParsers
-
 
 trait ParserTest extends JavaTokenParsers with Assertions {
